Function to report memory context stats of any backend
This function sends a signal to a backend to publish
statistics of all its memory contexts. Signal handler
sets a flag, which causes the relevant backend to copy its
MemoryContextStats to a fixed shared memory area, as part
of next CHECK_FOR_INTERRUPTS().
If there are more statistics to be shared, it creates
a file and copies remaining stats to that file.
Once its done, it signals the
client backend using a condition variable. The client backend
which was waiting on that condition variable, then wakes up,
reads the shared memory and returns these values in the form of
set of records, one for each memory context, to the user.
The client backend tries to read the remaining
statistics from the file if it exists. The client backend
is reponsible for deleting the file when it finishes
reading and also marking the shared memory area as not in
use, in order to allow the other backends to fill it with
new statistics.